# Work in progress: Protobuf to namedtuple