… .travis.yml
… README.md
… protozfits/CoreMessages_pb2.py
… protozfits/L0_pb2.py
… protozfits/VERSION
… protozfits/any_array_to_numpy.py
… protozfits/libACTLCore.dylib
… protozfits/libACTLCore.so
… protozfits/libZFitsIO.dylib
… protozfits/libZFitsIO.so
… protozfits/libprotobuf.9.dylib
… protozfits/libzmq.4.dylib
… protozfits/libzmq.so.3
… protozfits/patch_ids.py
… protozfits/rawzfitsreader.cpython-35m-darwin.so
… protozfits/rawzfitsreader.cpython-35m-x86_64-linux-gnu.so
… protozfits/rawzfitsreader.cpython-36m-darwin.so
… protozfits/rawzfitsreader.cpython-36m-x86_64-linux-gnu.so
… protozfits/simple.py
… protozfits/tests/resources/example_10evts.fits.fz
… protozfits/tests/test_rawzfitsreader.py
… protozfits/tests/test_read_raw_events.py
… protozfitsreader/VERSION
… protozfitsreader/__init__.py
… protozfitsreader/protobuf_to_namedtuple.py
… protozfitsreader/simple.py
… protozfitsreader/tests/test_rawzfitsreader.py
… protozfitsreader/tests/test_read_raw_events.py
… setup.py