Regarding issue #14 this is my first proposal:

 - [ ] missing: example notebook.

But in case you want to have a look ... go ahead.