@pep8speaks 2018-02-15
Hello @dneise! Thanks for submitting the PR.

 - In the file [`protozfitsreader/protobuf_to_namedtuple.py`](https://github.com/cta-sst-1m/protozfitsreader/blob/a30d7eeb27d48de1c3f6d908954ba7efcdc6a982/protozfitsreader/protobuf_to_namedtuple.py), following are the PEP8 issues :

> [Line 177:1](https://github.com/cta-sst-1m/protozfitsreader/blob/a30d7eeb27d48de1c3f6d908954ba7efcdc6a982/protozfitsreader/protobuf_to_namedtuple.py#L177): [E305](https://duckduckgo.com/?q=pep8%20E305) expected 2 blank lines after class or function definition, found 1