@dneise 2018-02-22
https://asciinema.org/a/8ejvuX6xTrvPx9RaiJ5Q9iPb0?speed=5