@dneise 2018-02-23
Okay ... so I just made a video .. where I explain how this PRs version compares to the current `ZFile`.
In that video I was showing some performance measurements, and to my own surprise, this PRs version was faster than the old version, although this PRs version of the code does more .. and doing more in less time is somehow strange.

Now I repeated the measurements I've done in the video on a larger file. And Now I get the expected result. The new version is slower ... it adds about 570µs per event, (that means, when only reading an event and not analysing it at all, this version here is 30% slower).

Here are the measrurements:
```python
from protozfits import ZFile  # old version
from protozfits.digicam import event_source  # new version

# old version
# measuring pure read time for 12509 events:
%time for e in ZFile('SST1M01_20171030.022.fits.fz'): pass
# results of several tries
CPU times: user 23 s, sys: 184 ms, total: 23.2 s
Wall time: 23.2 s
CPU times: user 23 s, sys: 180 ms, total: 23.1 s
Wall time: 23.1 s

# new version
# measuring pure read time for 12509 events:
%time for e in tqdm(event_source('SST1M01_20171030.022.fits.fz')): pass
# results of several tries
CPU times: user 30.2 s, sys: 88 ms, total: 30.3 s
Wall time: 30.3 s
CPU times: user 30 s, sys: 79.6 ms, total: 30 s
Wall time: 30 s
CPU times: user 30.2 s, sys: 87.9 ms, total: 30.2 s
Wall time: 30.2 s
```