@dneise 2018-02-26
np.argsort might take some time, and is now only done once on startup ... I did not profile this, so I do not know how much it improves anything