@dneise 2018-02-26
Ah .. maybe I should rename these files .. "simple" and "digicam" somehow before merging.