@dneise 2018-03-03
Oookay .. so I like this now a bit better .. I was able to remove some duplication from the code and tidy it up. What was know as `from protozfitsreader import ZFile` until now, is now called ` from protozfits import DigicamFile` because it is special for digicam.

For everybody else, they can use `from protozfits import SimpleFile` which is a bit slower than `DigicamFile` since it does more. 

Still faster than both of them is `SimpleFile(path, pure_protobuf=True)` ... it returns the same stuff, which `rawzfitsreader` is returning. So one still has to convert the `AnyArray`s to numpy arrays using the function: 
`from protozfits import any_array_to_numpy`

This is mostly described in the [README](https://github.com/cta-sst-1m/protozfitsreader/blob/protobuf_to_namedtuple/README.md) now as well. Please have a look.

Just to be clear, this branch is not compatible with digicampipe. The import in `digicampipe.io.zfits` needs to be changed. But we can still merge it, since digicampipe uses the protozfitsreader v0.44.2.

I hope this version makes playing around with zfits files easier for people from other telescopes.